A lot of changing, Fixing Shit
Ice Box use it own code so I can do more with it later on and Stone FirePit work and it won't cook things like ore but it will cook faster :D need to be low a bit. Added a GitIgnore so it won't upload dll and pdb, json in mods file. Not src file.